rename App to Resume. added a folder for myComponents
