+++ dreamjob/src/index.js
@@ -1,8 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './semantic.min.css';
-import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+// Import css
+import './semantic.min.css';
+
+// Import Components
+import Resume from './myComponents/Resume';
+
+ReactDOM.render(<Resume />, document.getElementById('root'));
 registerServiceWorker();
